08-28 6:27
골드방 이미지 변경, UI변경, 아이템 이미지 추가, 애니메이션 수정, 포트폴리오기간 1달차

+++ Isaac/Isaac/Tile.cpp
@@ -58,6 +58,6 @@ void Tile::LoadActor(std::wistringstream& steam)
 	std::string spriteName;
 	spriteName.assign(wspriteName.begin(), wspriteName.end());
 
-	_pos = Vector(x, y);
+	_pos = Vector((float)x, (float)y);
 	_sprite = CreateSpriteComponent(spriteName, w, h);
 }
